DatabaseAgent/requirements.txt
openai==1.58.1
python-dotenv==1.0.1
psycopg2==2.9.10